[rank] move scip utility to use in store
